Attach initiator frame routing id to renderer initiated navigations
This change routes the initiator_routing_id for the frame that initiates
a navigation to the NavigationRequest. This is only set on
a subset of navigations as some renderer initiated navigations are not
associated with an initiating frame, such as drag-and-drop navs.

The initiator RenderFrameHost is not guaranteed to exist by the time the
navigation is processed by the browser process because the frame's
deletion may have been scheduled prior to the navigation.

For this reason, the initiator_frame does not replace
|CommonNavigationParams.initiator_origin|.

This routing id is used to verify the top-level origin that an
impression occurred on for the Conversion Measurement API, rather than
relying on sending the origin from the renderer. This usage is
also implemented in this change.

BUg: 1059959
Change-Id: I8ab6d443a864cd72447d50a7d5a0588b68a19c61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090095
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765315}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f43556d014bc13d4f929d1683a1f8c84ad973524