Beforematch: Ensure to update style and layout in find-in-page.
This patch ensures that when we unlock the element in the beforematch
event, we still update style and layout.

The activate function is only used for bookkeeping and I will be
removing it, so we should not be relying on its return value in
order to decide whether we need an up-to-date layout.

R=jarhar@chromium.org, chrishtr@chromium.org

Change-Id: Iffbaae5229bddc78e41b4d0aa1cc15c4a3ad6429
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180115
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765321}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b3270e38db1e7bc1de2611bd040d1c913b4875bf