Early ignore mixed content CSP report-uri endpoints
Previously, any endpoints specified in the 'report-uri'
Content-Security-Policy directive which violate Mixed Content policy
would make the fetch throw an error on reporting. This was a
not-so-nice experience for the developer. With this change, we ignore
those endpoints already when parsing and we log a warning to the
console.

Bug: 695363
Change-Id: I8c0bea7525d199be88b066b1cd671280ec7d036a
Fixed: 695363
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166170
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765529}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8121919b9a819d2e455ae885f8dffa6d7c52c872