[EventTiming] Add a test for auxclick
This CL adds a helper function to add a test for an event type. The test
will add two fast events of the event type and one slow one and test
that the eventCounts as well as the entry in the PerformanceObserver for
the slow event.

Bug: 543598
Change-Id: I37da66424e7ea353b99787c578b26e3322cd984e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181147
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765537}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2202100baf8cf7d9312a114d56be51353e4d88c6