[blink scheduler] Next allowed run time is the earliest time aligned with the wake up interval.
Currently, if the Next Allowed Run Time is obtained from a
WakeUpBudgetPool when:

a) There is no last wake up and the Desired Run Time is aligned with
   the wake up interval -> The Next Allowed Run Time is
   Desired Run Time + Wake Up Interval.

b) The Desired Run Time is after (Last Wake Up + Wake Up Duration) and
   is aligned with the Wake Up Interval -> The Next Allowed Run Time is
   Desired Run Time + Wake Up Interval.

In both cases, that means that when the Wake Up Interval is 1 second
and a 1 second timer is set, it may fire in 2 seconds. Similarly, if
the Wake Up Interval is 1 minute and a 1 minute timer is set, it may
fire in 2 minutes.

With this CL, the Next Allowed Run Time is changed to always be the
earliest time that is at of after the Desired Run Time, and is aligned
with the Wake Up Interval.

Bug: 1075553
Change-Id: Ief7bf26732a6091557ef3c56588758bc181c70e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2175311
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765675}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ac39a205f6358e68c047778ec9f1a834312185d