[client-hints] Turn `Sec-CH-UA` into a parameterized list
The user-agent client hints spec was changed to a require
making the Sec-CH-UA "brands" header into a structured header
parameterized list. This will allow for adding the codebase a
project is based on (i.e. the various Chromium-based browsers).

Mostly just changing blink::UserAgentMetadata to use a list and
following down all the places it touches. Also fixed client hint
serialization in client_hints.cc, because why not, and centralized
the Sec-CH-UA serialization to the the UserAgentMetadata class. The
other header serialization can probably be centralized to the same
place but that can be for another CL.

Bug: 1070919, 1074326
Change-Id: I6dc74244d6c48bf2daed897bc1b4c13dd03e2248
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163260
Commit-Queue: Aaron Tagliaboschi <aarontag@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765745}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9f01b682dd5d376e5b80f1e1e7d114192ea045fb