Attaches event intents to deferred tree updates
Events that occur due to a tree update, such as a selection change
or a text change, and which cannot be fired at that moment
because the document lifecycle is not in the accessibility state
should remember their intents (i.e. what caused them) and
attach those intents, together with any active intents, to
the event when it actually gets fired in the accessibility lifecycle stage.
The source of the event (AKA event_from), such as "user" or "page", should also
be remembered, because otherwise a user-initiated action for example will
not get the correct "event_from" when the corresponding event has been deferred.

AX_RelNotes: N/A

R=dmazzoni@chromium.org

Change-Id: Ia15113600e2083c203e0da321a63d1f2a1342908
Bug: 989156
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181673
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765752}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa94014dbe0292387e035013a22cfe0e16d8f701