Properly update Unadjusted Movement in |PointerLockController|
Previously the |current_unadjusted_movement_setting_| was only updated
when a Request was made.  It was not updated if that request failed or
if a change request was made. Now it updates upon successful lock or
change.

Bug: 1073439
Change-Id: Ifabbbeedefd9035d334ce0832ebaebadf079f3e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180931
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: James Hollyer <jameshollyer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765769}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90eb1b6f8b81ae3f8e6adb82950782757ba05012