Update android_deps to TOT.
Some files need to be updated to the latest, as these only get
updated if there is a change in version. Most changes are just
a few rewords or removing a newline character.

The only significant change is
com_google_code_findbugs_jsr305/README.chromium is updated to
shipped and secuirty critical. This was missed because we moved the
dependencies from third_party/findbugs to android_deps/findbugs (both
libraries were included) but the fetch_all.py script only updates the
README if there is a version change, it didn't regenerate it  when
findbugs was added as a compile dependency in
https://chromium-review.googlesource.com/c/chromium/src/+/2079436/9/third_party/android_deps/build.gradle

Add bug to address in future.
https://bugs.chromium.org/p/chromium/issues/detail?id=1078529

Bug: 1067050
Change-Id: I80e9f2a7d54723d1c9982710919c66bb95025b7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183072
Commit-Queue: benjamin joyce <bjoyce@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765865}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2126f2741752cde9b811ee175f6e4aac88415353