appcache: add feature for gating AppCache on origin trials
The state of features with this patch is:

(1) kAppCache feature

Enabled by default until we finally turn off AppCache just prior to
removal.

If this is disabled, then AppCache is forced disabled everywhere and
will not be turned on, even by the origin trial.

If this is enabled, then the AppCache is enabled, but access may be
gated on the origin trial if kAppCacheOriginTrial is enabled.

This can currently be set by about:flags, and from WebView for testing
purposes.

(2) kAppCacheRequireOriginTrial feature (new)

Disabled by default.  This will switch to enabled by default,
most likely in M85.  Followup patches will add settings to enterprise
and WebView to disable this feature optionally.

If enabled, then it will gate the API on origin trial tokens in meta
tags / http headers, and will gate the browser code on origin trial
tokens in manifests.

Bug: 582750
Change-Id: I4d87c702d419030cd3f382a2c2796ce406a7437e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142615
Commit-Queue: enne <enne@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766086}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b53b437fa7ec9bc6e9fc8bb24b770a7ebf58c218