Reland "heap: Remove GetHeapObjectHeader from mixins"
The methods in the macro are put on every GarbageCollectedMixin which
is expensive in terms of binary size. The HoH getter can be replaced
with GetTraceDescriptor().

The reland adds a proper bottleneck for retrieving the object header
that handles the in-construction case.

This reverts commit 0897b1c3ee067334934045692ca4689548badfb6.

Bug: chromium:1071874
Change-Id: I95c2e01ac1348be2d5acbc70fea07e80da16e6f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153082
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766135}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9f90837c85ba58fcd5020616361a1ecae50ae67d