<select>: Fix a DCHECK failure on adding a not-owned OPTION element
When an OPTION element is added to an OPTGROUP element,
HTMLOptGroupElement::ChildrenChanged() is responsible to notify it to
the OPTGROUP's owner SELECT element. The function checks existence of
the owner SELECT first, but it failed because of the wrong
implementation of HTMLOptGroupElement::OwnerSelectElement().

This CL fixes OwnerSelectElement().

Bug: 1077556
Change-Id: If1bd4dbd022954c6caa55a8e44d3a31e2994b55b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186786
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Auto-Submit: Kent Tamura <tkent@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766341}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2e67a5cdd2fc62fafb903f1be1348146f2fc059e