[Security] Handle speculative RFHs better with COOP.
This patch addresses some inconsistencies with the way speculative RFH
and COOP interact.

Currently:
- We never switch for a speculative RFH, and don't have any COOP value.
- Anything that never has a network COOP never triggers a swap.

This is awkward because implementing the process model relies on having
homogeneous COOP for pages within an isolated BrowsingInstance.
If we base the process model on what we currently have, we end up with
all the speculative and non-network pages being put in a potentially
crossOriginIsolated privileged process.

After this patch:
- (Almost) all speculative and non-network pages are computing if they
need a COOP swap.
- We exclude "about:blank" from the computation, because the current
test_runner infrastructure hangs with a process swap on the cleanup
page.
- We inherit the "same-origin" value of the current page. This is
required because during a crash, an optimization forces the speculative
RFH to be committed instantly. At reload time, we can stay in the same
BrowsingInstance.
- We currently inherit same-origin-allow-popups only when we are not
just opening a popup. This is otherwise incompatible from the way we
store "if we swapped" on the NavigationRequest. In the scenario where a
page "a.com" with same-origin-allow-popups opens a popup to "b.com"
with no coop, the inherit would cause a swap (that is remembered in
the NavigationRequest) when there ends up being none. This "memory"
mechanism is necessary for redirects and also used by name clearing.


Change-Id: I1b9116c66e19c8f5c2ccda2f9c42e4085e8f0bc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2162751
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766390}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 738ba65520ed0c44d0034ae511e552500e179e56