Improve test: csp/reporting/post-redirect-stacktrace.https.html
The test was checking no post-redirect information was leaked.
Unfortunately, there was a bug in the test. The same data was present in
both the final URL and the post-redirect URL.

This patch make the post-redirect URL no to have the data.

BUG=1074317

Change-Id: I40a698b8edf6ff651dda5d0e82b803a60ccaac7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187711
Reviewed-by: Daniel Vogelheim <vogelheim@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f4152597fb034b9c520eebdaa6902e7673ff453d