Removing kSoftwareCompositedUsage for CanvasRenderingContextHost in 3D
It is not needed to try and use the SharedImageProvider as currently
the CanvasResourceProvider::Create method will instantly fall into the
SharedBitmapProvider fallback or BitmapProvider fallback.

This CL removes the SoftwareCompositedUsage and replaces it with both
fallbacks.

This CL also adds the CreateSharedBitmapProvider.

Bug: 1035589
Change-Id: I2cc0b23c1a40c15285ccdbba06840a0f98ff1b39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2118573
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766425}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 788b2ea2f5dfdc8ce5edf66ba00835dc773291b6