[devtools] Add filename to Page.downloadWillBegin
Chromium (and browsers in general) have logic about how to name a
downloaded file based on a number of factors: `Content-Disposition`
HTTP headers, the `download` attribute of an Anchor tag, detected
MimeType, etc. When testing and understanding what the browser
downloaded, it's helpful to know what the download was named on
disk.

Previously, there was no way to tie a Download GUID (which appears in
the `Page.downloadWillBegin` and `Page.downloadProgress` events) to the
actual filename on disk.

This changeset fixes that by exposing an approximated filename in the
Page.downloadWillBegin event.

NB: This does not fully encompass all the browser-level renaming
possibilities, but it is a near approximation of what is likely to
end up on disk. Certain extensions and mimetypes may yield different
final names; likewise, downloaded files with clashing names will not
include the incrementing counter number that users would see in Chrome
when downloading files with clashing names.

Test: Ran the following targets: `http/tests/inspector-protocol/{browser,page}`
Bug: chromium:1076697
Change-Id: I48e364b1696627ef19ec25049838f88d085aba2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174778
Reviewed-by: Pavel Feldman <pfeldman@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Pavel Feldman <pfeldman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766623}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0900dee8e2001c70da70f446a717e99430e87e39