[LayoutNG] Add URLs in ng_box_fragment_painter
The AddPDFURLRectsForInlineChildrenRecursively method was never
implemented for LayoutNG. This results in missing URLs for PaintPreviews
(this may also apply to print-to-PDF although I haven't been able to
confirm this).

To correct this AddPDFURLRectsForInlineChildrenRecursively needs to be
implemented.

Bug: 1079306
Change-Id: I38ad0860c458916d848f4b61c8d4cecc46d873d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182944
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766842}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f52b34bef67e4189995dcd396c26f7792f4aedb8