Add visibility in PageLifecycleStateManager
This CL adds visibility in PageLifecycleStateManager and replaces
SetVisibility page message.

PageLifecycleStateManager manages the page's freezing states and the
visibility states, and sends changes to the renderer.

Context here is that, PageLifecycleStateManager was introduced to
maintain the freezing/visibility status on the browser side in a single
point. Before this, there was no single place to control freezing/
visibility state and anybody could change them by calling IPCs.

PageLifecycleStateManager is the single entry point for changing the
freezing/visibility states and is responsible for sending updates to
renderer. This is necessary to solve nested freezing/visibility.

Bug: 1057506
Change-Id: I351b146c7c86882dd6b7fbcfd44e484a4de487a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100171
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Yuzu Saijo <yuzus@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#767207}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be024af0dbe684b85a7a4e80ee53b13228a40f27