Don't load CSS resources for non-rendered elements.
The current situation was that we triggered loading of all resources,
like background-images, for an element we computed style for. Instead,
align with Firefox and Safari to skip fetching for:

* display:none elements
* elements in display:none subtrees
* display:contents elements
* elements outside the flat tree

Note that we still trigger fetches for visibility:hidden, which is in
line with what Firefox and Safari does.

Also removed an unnecessary StyleResolver::LoadPendingResources wrapper.

Bug: 765095
Change-Id: I958b5d22bac942408957398bf5761269908466b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190650
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#767327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: caf740f6c0b57ca102026f40f3cc62f9f3cf5159