Add new canvas API flag
The canvas team is going to implement new features for the api:
https://github.com/fserb/canvas2d

Create a runtime flag to put these features behind for now. We're going
to need this flag for at least a year, so I added it to the never
expire list.

Bug: 1078565
Change-Id: I80df0d646dbc9c08681f687c5a65d4422c75b249
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183133
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#767510}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 229cffd082ba390c0b62f127c10bf6ed81110ba0