[FlexNG] Mark containing-block chain as maybe have %-height descendants.
This change marks the containing-block chain of an object as maybe
containing %-height-descendants when
LayoutBlock::AddPercentHeightDescendant is called.

This allows the NG caching logic to skip layout for whenever there is
a legacy layout object root in the subtree, and either the %-block-size
or definiteness changes.

Bug: 845235
Change-Id: I10e7b0c35707cc668e92e23ed690a50ee4f9a02d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194346
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#767545}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e35064ad0a4bc80afcbe89b8880226d09823a6c8