[css-pseudo] Reuse the CSSPropertyValueSet with the UA ::marker styles
Since r2182427, a new MutableCSSPropertyValueSet object was created when
resolving the styles of a ::marker pseudo-element. This caused a memory
regression for sites with lots of list items (though it was not that bad
since the objects were garbage collectible).

This patch fixes it by making a persistent MutableCSSPropertyValueSet
and reusing it for all markers.

BUG=1081143

Change-Id: If5a8e0bd6928811c2759dc4891e36e71ab150a6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193777
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#767763}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dec4988ca59ce4b712c6eff368b890377d4bfa80