Remove use of toTitleCase in layout tests
It's a method that's added to the String prototype by DevTools frontend,
but we are removing that in a CL [1]. The layout tests need to be
updated to not depend on it. They don't need the expression to be in
title case so we'll just change the tests to use the expression as is.

[1]: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2193590

Change-Id: I7ac36003423ded30ec42072a0667d66af590215c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2195962
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#767773}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62a53971766cae58654cfeb5d766f75b0891d5af