WebXR - AR - anchors - bring anchor creation up to spec
With the latest spec change, WebXR Anchors API should strive to create
anchors such that the anchor location reflects the knowledge of the
world (& the spaces) as of the time of XRFrame used to make the call.
For Chrome's implementation, it means that we should rely of the notion
of stationary & dynamic XRSpaces, and in case the app wants to create
an anchor relative to a space that is considered a dynamic one, we
should use the current knowledge of the world to calculate anchor's pose
w.r.t. a stationary XRSpace that is deemed best - currently, we will
attempt to use local space for anchor creation and fall back to
unbounded reference space.

Spec draft PR with the corresponding change to the spec:
https://github.com/immersive-web/anchors/pull/42

Change-Id: I25398844cd7d92dc8a2af76c8175e6098b099452
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194594
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768007}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de5d2b840a46d057425d9ff605de692f824d1e84