[FragmentItem] Implement line cache
This patch implements line cache for |NGFragmentItems|. The
line cache keeps track of which line was modified. When
running re-layouts, we reuse lines up to the first modified
lines. This improves when appending to a large text, or when
parts of a large text was modified.

This is similar to |LayoutNGLineCache|, our last attempt to
implement line cache for LayoutNG, but differs in:
1. Implements on top of |NGFragmentItem|.
2. Shares most of cache validity logic with simplified
   layout.

Having two complicated but slightly different cache validity
logic was the biggest concern in the previous attempt. This
patch solves the concern.

This improves:
- line-layout-repeat-append by ~60%.
- line-layout-repeat-append-select by ~36%.
- paint/appending-text.html does not improve, because it is a
  large text of single line.
https://pinpoint-dot-chromeperf.appspot.com/job/14945cf4120000
https://pinpoint-dot-chromeperf.appspot.com/job/15eea6afe20000

The old |LayoutNGLineCache| is not removed yet, as it
provides good reference for the new line cache. I will remove
it once this line cache is stabilized.

Bug: 982194
Change-Id: I573e52aa020057e14aa90baf1f06c55c40e4a1f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185249
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768067}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b7bc9cdd8ad0b2088806b191497e495f89e42ec