Adding new option that includes webGPU flag in CreateSharedImageProvider
There was a small thing missing in the CL
https://chromium-review.googlesource.com/c/chromium/src/+/2171319
that focus on Implementing the Dawn Backend for CanvasResourceProvider.

In the next couple of weeks, the method of
CanvasResourceProvider::Create and the fallback logic will be removed,
and only the CanvasResourceProvider::CreateSharedImageProvider
method will be used to create SharedImageProviders, in the above CL
is_webgpu was set to false, always ignoring the DAWN backend.
This method actually was already used in several places where Create
was already deleted.

This CL adds that flag to the CreateSharedImageProvider, setting
dawn mode if the feature is enabled in canvas.


Bug: 1076141, 1035589
Change-Id: Ie8f8f906d7375a4408133310526c84274dd63669
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190839
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Stephen White <senorblanco@chromium.org>
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768085}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 407e8fba907a411c9944db1a523aaacac7889506