Preserve LCD-text in SaveLayerAlphaOp when possible
This achieves LCD-text under non-composited opacity when possible. Pass
two flags to SkCanvas::SaveLayer() for SaveLayerAlphaOp:
kPreserveLCDText_SaveLayerFlag to preserve LCD text, and
kInitWithPrevious_SaveLayerFlag to to ensure an opaque background.

This doesn't apply when there any SaveLayer modes which would break
LCD text or be broken by kInitWithPrevious_SaveLayerFlag. So we disable
the feature for the whole layer when there is any of such effects in
the layer. The new operation is also a bit more expensive, so only do
it when there is text. For now these two conditions are applied on the
whole layer for simplicity. Otherwise we would need to track them in
SaveLayerAlpha/Restore hierarchy which would be complex and error-prone.

Bug: 942051
Change-Id: I3ab2e8051e26c5714fd72830a7921099b9170d0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189260
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768367}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0c258cbd3a2a4e8d1ebf82db8489672ec944f1a7