[ScrollUnification] Only collect scroll translations behind feature flag
In https://crrev.com/c/2089973, the ScrollUnification feature flag was
added along with various changes to ensure the compositor has the
complete scroll tree. Most of that CL was behind the feature flag,
except for traversing the LocalFrameView set to gather all scroll
translation nodes, which is likely the cause of the regressions in
bug 1081923. This patch puts this step behind the feature flag as well.

Bug: 1081923 1047178
Change-Id: I6ee3e3f60c29c1739c3221871e0baf8a96d114ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199597
Commit-Queue: Nick Burris <nburris@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768518}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6059c240e9ca4aceac9563e53bedb64c0bf6c3a6